+++ src/screens/recipient/DetailedGoalCard.tsx
@@ -4,7 +4,6 @@ import {
   Text,
   TouchableOpacity,
   StyleSheet,
-  Alert,
   Animated,
   LayoutAnimation,
   Pressable,